004-content-personalization/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
python-multipart
pyjwt
httpx
psycopg2-binary